Rename to flatGeojsonToBinary
