+++ modules/gis/src/index.ts
@@ -1,7 +1,7 @@
 // Types from `@loaders.gl/schema`
 
 // Functions
-export {featuresToBinary} from './lib/features-to-binary';
+export {flatGeojsonToBinary} from './lib/flatGeojson-to-binary';
 export {geojsonToBinary, geojsonToBinaryOld} from './lib/geojson-to-binary';
 export {geojsonToFlatGeojson} from './lib/geojson-to-flatGeojson';
 export {binaryToGeojson, binaryToGeoJson, binaryToGeometry} from './lib/binary-to-geojson';
