+++ modules/gis/src/lib/flatGeojson-to-binary.ts
